% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateSkeleton.R
\name{generateSkeleton}
\alias{generateSkeleton}
\title{Title}
\usage{
generateSkeleton()
}
\value{
folder structure, app.R file, and global.R file
}
\description{
Title
}
\examples{
generateSkeleton()
}
